restructure template rendering
