+++ app/src/displays/related-values/index.ts
@@ -1,6 +1,6 @@
 import { defineDisplay } from '@/displays/define';
 import DisplayRelatedValues from './related-values.vue';
-import getFieldsFromTemplate from '@/utils/get-fields-from-template';
+import { getFieldsFromTemplate } from '@/utils/render-template';
 import adjustFieldsForDisplays from '@/utils/adjust-fields-for-displays';
 import getRelatedCollection from '@/utils/get-related-collection';
 import useCollection from '@/composables/use-collection';
