+++ app/src/interfaces/repeater/repeater-row-header.vue
@@ -11,7 +11,7 @@
 import { defineComponent, computed } from '@vue/composition-api';
 import { render } from 'micromustache';
 import i18n from '@/lang';
-import getFieldsFromTemplate from '@/utils/get-fields-from-template';
+import { getFieldsFromTemplate } from '@/utils/render-template';
 
 export default defineComponent({
 	props: {
