--- app/src/utils/get-fields-from-template/get-fields-from-template.ts
@@ -1,13 +1,0 @@
-export default function getFieldsFromTemplate(string: string) {
-	const regex = /{{(.*?)}}/g;
-	let fields = string.match(regex);
-
-	if (!Array.isArray(fields)) {
-		return [];
-	}
-
-	fields = fields.map((field) => {
-		return field.replace(/{{/g, '').replace(/}}/g, '').trim();
-	});
-	return fields;
-}
