--- app/src/utils/get-fields-from-template/index.ts
@@ -1,4 +1,0 @@
-import getFieldsFromTemplate from './get-fields-from-template';
-
-export { getFieldsFromTemplate };
-export default getFieldsFromTemplate;
