+++ app/src/utils/render-template/index.ts
@@ -1,0 +1,4 @@
+import { getFieldsFromTemplate, renderTemplate } from './render-template';
+
+export { getFieldsFromTemplate, renderTemplate };
+export default renderTemplate;
